--- src/_base/_decorators/model-name.decorator.ts
@@ -1,4 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-
-export const ModelName = (modelName: string) =>
-  SetMetadata('modelName', modelName);
